hooked up to a backend

+++ src/mortgage.amortization.api/Api/Amortization/CalculateAmortizationRequest.cs
@@ -11,7 +11,7 @@ namespace mortgage.amortization.api.Api.Amortization
         [JsonPropertyName("loan")]
         public decimal Loan { get; set; }
         [JsonPropertyName("period")]
-        public decimal Period { get; set; }
+        public int Period { get; set; }
         [JsonPropertyName("rate")]
         public decimal Rate { get; set; }
     }
